richard-api/pyproject.toml
[project]
name = "richard-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "google-auth>=2.28.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "boto3>=1.40.26",
    "youtube-transcript-api>=1.2.2",
    "openai>=1.107.0",
    "yt-dlp>=2025.9.5",
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.13",
]
